Fixed the transition from seqplayer to pattern generator regarding the ZMP and the heuristic for the arms.
There is still an asymmetry regarding the generated angle which could be
explained by the filter smoothing the reference ZMP, also the motion
asked is suppose to be symmetric (stepping).

It works on OpenHRP-2 dynamic simulation, but with a slight
overshoot in the force on one of the feet in stepping phase.
It has to be confirmed by experience if this is a problem or not,
and by comparison with Kajita's walkfwd.

Thus this commit is still classified "dev only."