Computation of the Specialized Inverted Kinematics for Kajita's method.
In ComAndFootRealizationByGeometry the computation are now done
using the new abstract interface, and the optimized implementation.
Propagated the modification across the framework.
For this commit, the system is still not working fully perfectly,
but ZMPDiscretization seems to behave correctly.
Corrected a problem in the Filter of the ZMPReference.
Changed some licenses reference still not fully correct.