Removed documentation file.
Following complaints from users on the downloading time,
those files are removed. Thus it will still be long to
clone the repository but no extension of the time is
to be expected if the documentation is changed.
Moreover the next documentation will be limited
to the public interface, generated by doxygen.
For people interested by the algorithms, please
read the papers given in reference in each source file.