Fix precision problem.
The float variables introduced a precision problem
during time intervals computation. This lead to failures when testing limit
condition such start and end of the intervals.
Indeed the classical parameters such 0.78 for the single support phase
and 0.02 for the double support DO NOT have an exact representation
in the space of floats.
Every variables of ZMPRefTrajectoryGeneration are now double
which allow exact representation for the golden parameters.