Test size of variables before trying to allocate memory.
When trying an array of data one should test if the size is zero.
Indeed in such case, the time to handle the error by the system can be prohibitive.

This commit lower the time to allocate memory from 18 ms to 8 ms.

A better policy would be to allocate memory when called by the control
part which is not supposed to in the real-time part of the controller.