Fix ZMP computation for multibody correction.
With the use of spatial algebra in JRL-dynamics ZMP needs
backward dynamics computation for proper evaluation.
The internal setting of the WPG used an approximation of
the momentum derivatives previously implemented in JRL-dynamics.
This commit fix this problem and the references for the unitary
test.

This should be also validate with the simple-humanoid and not
HRP-2.